+++ src/commands/admin.h
@@ -3,6 +3,8 @@
 
 __SWIRC_BEGIN_DECLS
 void cmd_die(const char *);
+void cmd_gline(const char *);
+void cmd_kline(const char *);
 void cmd_rehash(const char *);
 void cmd_restart(const char *);
 void cmd_wallops(const char *);