+++ src/commands/info.h
@@ -5,6 +5,7 @@ __SWIRC_BEGIN_DECLS
 void cmd_admin(CSTRING);
 void cmd_info(CSTRING);
 void cmd_ison(CSTRING);
+void cmd_servstats(CSTRING);
 __SWIRC_END_DECLS
 
 #endif