+++ src/readline.h
@@ -47,9 +47,11 @@ typedef struct tagREADLINE_POS {
 typedef struct tagTAB_COMPLETION {
 	char	search_var[64];
 	bool	isInCirculationModeForConnect;
+	bool	isInCirculationModeForCs;
 	bool	isInCirculationModeForHelp;
 	bool	isInCirculationModeForMsg;
 	bool	isInCirculationModeForNotice;
+	bool	isInCirculationModeForNs;
 	bool	isInCirculationModeForQuery;
 	bool	isInCirculationModeForSasl;
 	bool	isInCirculationModeForSettings;