+++ src/readlineTabCompletion.c
@@ -31,6 +31,7 @@
 
 #include "commands/connect.h"
 #include "commands/sasl.h"
+#include "commands/services.h"
 #include "commands/squery.h"
 #include "commands/theme.h"
 #include "commands/znc.h"
@@ -72,6 +73,13 @@ auto_complete_connect(volatile struct readline_session_context *ctx,
 }
 
 static void
+auto_complete_cs(volatile struct readline_session_context *ctx,
+    CSTRING s)
+{
+	do_work(ctx, L"/cs ", s);
+}
+
+static void
 auto_complete_help(volatile struct readline_session_context *ctx,
     CSTRING s)
 {
@@ -90,6 +98,13 @@ auto_complete_notice(volatile struct readline_session_context *ctx,
     CSTRING s)
 {
 	do_work(ctx, L"/notice ", s);
+}
+
+static void
+auto_complete_ns(volatile struct readline_session_context *ctx,
+    CSTRING s)
+{
+	do_work(ctx, L"/ns ", s);
 }
 
 static void
@@ -243,9 +258,11 @@ readline_tab_comp_ctx_new(void)
 	BZERO(ctx.search_var, sizeof ctx.search_var);
 
 	ctx.isInCirculationModeForConnect	= false;
+	ctx.isInCirculationModeForCs		= false;
 	ctx.isInCirculationModeForHelp		= false;
 	ctx.isInCirculationModeForMsg		= false;
 	ctx.isInCirculationModeForNotice	= false;
+	ctx.isInCirculationModeForNs		= false;
 	ctx.isInCirculationModeForQuery		= false;
 	ctx.isInCirculationModeForSasl		= false;
 	ctx.isInCirculationModeForSettings	= false;
@@ -278,9 +295,11 @@ readline_tab_comp_ctx_reset(PTAB_COMPLETION ctx)
 		BZERO(ctx->search_var, sizeof ctx->search_var);
 
 		ctx->isInCirculationModeForConnect	= false;
+		ctx->isInCirculationModeForCs		= false;
 		ctx->isInCirculationModeForHelp		= false;
 		ctx->isInCirculationModeForMsg		= false;
 		ctx->isInCirculationModeForNotice	= false;
+		ctx->isInCirculationModeForNs		= false;
 		ctx->isInCirculationModeForQuery	= false;
 		ctx->isInCirculationModeForSasl		= false;
 		ctx->isInCirculationModeForSettings	= false;
@@ -314,6 +333,22 @@ init_mode_for_connect(volatile struct readline_session_context *ctx)
 	ctx->tc->elmt = textBuf_head(ctx->tc->matches);
 	auto_complete_connect(ctx, ctx->tc->elmt->text);
 	ctx->tc->isInCirculationModeForConnect = true;
+}
+
+static void
+init_mode_for_cs(volatile struct readline_session_context *ctx,
+    const int offset)
+{
+	char	*p = addrof(ctx->tc->search_var[offset]);
+
+	if ((ctx->tc->matches = get_list_of_matching_cs_cmds(p)) == NULL) {
+		output_error("no magic");
+		return;
+	}
+
+	ctx->tc->elmt = textBuf_head(ctx->tc->matches);
+	auto_complete_cs(ctx, ctx->tc->elmt->text);
+	ctx->tc->isInCirculationModeForCs = true;
 }
 
 static void
@@ -359,6 +394,22 @@ init_mode_for_notice(volatile struct readline_session_context *ctx)
 	ctx->tc->elmt = textBuf_head(ctx->tc->matches);
 	auto_complete_notice(ctx, ctx->tc->elmt->text);
 	ctx->tc->isInCirculationModeForNotice = true;
+}
+
+static void
+init_mode_for_ns(volatile struct readline_session_context *ctx,
+    const int offset)
+{
+	char	*p = addrof(ctx->tc->search_var[offset]);
+
+	if ((ctx->tc->matches = get_list_of_matching_ns_cmds(p)) == NULL) {
+		output_error("no magic");
+		return;
+	}
+
+	ctx->tc->elmt = textBuf_head(ctx->tc->matches);
+	auto_complete_ns(ctx, ctx->tc->elmt->text);
+	ctx->tc->isInCirculationModeForNs = true;
 }
 
 static void
@@ -561,17 +612,47 @@ no_more_matches(volatile struct readline_session_context *ctx)
 	readline_tab_comp_ctx_reset(ctx->tc);
 }
 
+static bool
+var_matches_cs(CSTRING sv, int *iout)
+{
+	if (!strncmp(sv, "/chanserv ", 10))
+		*iout = 10;
+	else if (!strncmp(sv, "/cs ", 4))
+		*iout = 4;
+	else
+		*iout = 0;
+	return (!strncmp(sv, "/chanserv ", 10) || !strncmp(sv, "/cs ", 4));
+}
+
+static bool
+var_matches_ns(CSTRING sv, int *iout)
+{
+	if (!strncmp(sv, "/nickserv ", 10))
+		*iout = 10;
+	else if (!strncmp(sv, "/ns ", 4))
+		*iout = 4;
+	else
+		*iout = 0;
+	return (!strncmp(sv, "/nickserv ", 10) || !strncmp(sv, "/ns ", 4));
+}
+
 static void
 init_mode(volatile struct readline_session_context *ctx)
 {
+	int off1, off2;
+
 	if (!strncmp(get_search_var(ctx), "/connect ", 9))
 		init_mode_for_connect(ctx);
+	else if (var_matches_cs(get_search_var(ctx), &off1))
+		init_mode_for_cs(ctx, off1);
 	else if (!strncmp(get_search_var(ctx), "/help ", 6))
 		init_mode_for_help(ctx);
 	else if (!strncmp(get_search_var(ctx), "/msg ", 5))
 		init_mode_for_msg(ctx);
 	else if (!strncmp(get_search_var(ctx), "/notice ", 8))
 		init_mode_for_notice(ctx);
+	else if (var_matches_ns(get_search_var(ctx), &off2))
+		init_mode_for_ns(ctx, off2);
 	else if (!strncmp(get_search_var(ctx), "/query ", 7))
 		init_mode_for_query(ctx);
 	else if (!strncmp(get_search_var(ctx), "/sasl ", 6))
@@ -624,12 +705,16 @@ readline_handle_tab(volatile struct readline_session_context *ctx)
 		return;
 	} else if (ctx->tc->isInCirculationModeForConnect) {
 		ac_doit(auto_complete_connect, ctx);
+	} else if (ctx->tc->isInCirculationModeForCs) {
+		ac_doit(auto_complete_cs, ctx);
 	} else if (ctx->tc->isInCirculationModeForHelp) {
 		ac_doit(auto_complete_help, ctx);
 	} else if (ctx->tc->isInCirculationModeForMsg) {
 		ac_doit(auto_complete_msg, ctx);
 	} else if (ctx->tc->isInCirculationModeForNotice) {
 		ac_doit(auto_complete_notice, ctx);
+	} else if (ctx->tc->isInCirculationModeForNs) {
+		ac_doit(auto_complete_ns, ctx);
 	} else if (ctx->tc->isInCirculationModeForQuery) {
 		ac_doit(auto_complete_query, ctx);
 	} else if (ctx->tc->isInCirculationModeForSasl) {