+++ src/events/stats.h
@@ -1,0 +1,8 @@
+#ifndef GUARD_STATS_H
+#define GUARD_STATS_H
+
+__SWIRC_BEGIN_DECLS
+void event_statskline(struct irc_message_compo *);
+__SWIRC_END_DECLS
+
+#endif