+++ CHANGELOG.md
@@ -20,6 +20,8 @@ All notable changes to this project will be documented in this file.
   - 257 (`RPL_ADMINLOC1`)
   - 258 (`RPL_ADMINLOC2`)
   - 259 (`RPL_ADMINEMAIL`)
+  - 371 (`RPL_INFO`)
+  - 374 (`RPL_ENDOFINFO`)
 
 ## [3.4.4] - 2023-07-29 ##
 - **Added** a man page that describes how to write themes.