+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] ##
 - Added command-line option `-W`
 - Added tab completion for:
+  - `/cs`, `/chanserv`
+  - `/ns`, `/nickserv`
   - `/squery`
 - Added the following commands:
   - `/admin`